refactor(parser): remove legacy lexer modules
