+++ crates/assembly-syntax/src/parser/tests.rs
@@ -77,115 +77,6 @@ fn assert_parses(source: Arc<SourceFile>) {
     parse_forms(source).expect("parser should succeed");
 }
 
-// This test checks the lexer behavior with regard to tokenizing `exp(.u?[\d]+)?`
-#[test]
-fn lex_exp() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new("begin exp.u9 end");
-    let mut lexer = Lexer::new(source_id, scanner).map(|result| result.map(|(_, t, _)| t));
-    assert_matches!(lexer.next(), Some(Ok(Token::Begin)));
-    assert_matches!(lexer.next(), Some(Ok(Token::ExpU)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Int(n))) if n == 9);
-    assert_matches!(lexer.next(), Some(Ok(Token::End)));
-}
-
-#[test]
-fn lex_block() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new(
-        "\
-const ERR1 = 1
-
-begin
-    u32assertw
-    u32assertw.err=ERR1
-    u32assertw.err=2
-end
-",
-    );
-    let mut lexer = Lexer::new(source_id, scanner).map(|result| result.map(|(_, t, _)| t));
-    assert_matches!(lexer.next(), Some(Ok(Token::Const)));
-    assert_matches!(lexer.next(), Some(Ok(Token::ConstantIdent("ERR1"))));
-    assert_matches!(lexer.next(), Some(Ok(Token::Equal)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Int(1))));
-    assert_matches!(lexer.next(), Some(Ok(Token::Begin)));
-    assert_matches!(lexer.next(), Some(Ok(Token::U32Assertw)));
-    assert_matches!(lexer.next(), Some(Ok(Token::U32Assertw)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Dot)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Err)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Equal)));
-    assert_matches!(lexer.next(), Some(Ok(Token::ConstantIdent("ERR1"))));
-    assert_matches!(lexer.next(), Some(Ok(Token::U32Assertw)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Dot)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Err)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Equal)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Int(2))));
-    assert_matches!(lexer.next(), Some(Ok(Token::End)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Eof)));
-}
-
-#[test]
-fn lex_emit() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new(
-        "\
-begin
-    push.1
-    emit.event(\"abc\")
-end
-",
-    );
-    let mut lexer = Lexer::new(source_id, scanner).map(|result| result.map(|(_, t, _)| t));
-    assert_matches!(lexer.next(), Some(Ok(Token::Begin)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Push)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Dot)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Int(1))));
-    assert_matches!(lexer.next(), Some(Ok(Token::Emit)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Dot)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Event)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Lparen)));
-    assert_matches!(lexer.next(), Some(Ok(Token::QuotedIdent("abc"))));
-    assert_matches!(lexer.next(), Some(Ok(Token::Rparen)));
-    assert_matches!(lexer.next(), Some(Ok(Token::End)));
-    assert_matches!(lexer.next(), Some(Ok(Token::Eof)));
-}
-
-#[test]
-fn lex_invalid_token_after_whitespace_returns_error() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new("begin \u{0001}\nend\n");
-    let mut lexer = Lexer::new(source_id, scanner).map(|result| result.map(|(_, t, _)| t));
-
-    assert_matches!(lexer.next(), Some(Ok(Token::Begin)));
-    assert_matches!(
-        lexer.next(),
-        Some(Err(ParsingError::InvalidToken { span })) if span.into_range() == (6..7)
-    );
-}
-
-#[test]
-fn lex_invalid_underscore_token_span() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new("begin _-\nend\n");
-    let mut lexer = Lexer::new(source_id, scanner).map(|result| result.map(|(_, t, _)| t));
-
-    assert_matches!(lexer.next(), Some(Ok(Token::Begin)));
-    assert_matches!(
-        lexer.next(),
-        Some(Err(ParsingError::InvalidToken { span })) if span.into_range() == (6..7)
-    );
-}
-
-#[test]
-fn lex_single_char_token_and_ident_spans() {
-    let source_id = SourceId::default();
-    let scanner = Scanner::new("@\nA\n");
-    let mut lexer = Lexer::new(source_id, scanner);
-
-    assert_matches!(lexer.next(), Some(Ok((0, Token::At, 1))));
-    assert_matches!(lexer.next(), Some(Ok((2, Token::ConstantIdent("A"), 3))));
-}
-
 #[test]
 fn overlong_path_component_is_rejected_without_panic() {
     use std::panic::{AssertUnwindSafe, catch_unwind};
